CLN: Move GOES download functions to downloader.py
